chore: reduce wallet page routing complexity
