+++ components/brave_wallet_ui/components/desktop/views/index.ts
@@ -1,9 +1,16 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import { Account, Accounts } from './accounts'
 import CryptoView from './crypto'
-import PortfolioView from './portfolio'
-import AccountsView from './accounts'
+import { PortfolioAsset, PortfolioOverview } from './portfolio'
 
 export {
+  Account,
+  Accounts,
   CryptoView,
-  PortfolioView,
-  AccountsView
+  PortfolioAsset,
+  PortfolioOverview
 }
