+++ components/brave_wallet_ui/components/desktop/wallet-onboarding/welcome/index.tsx
@@ -1,5 +1,19 @@
-import * as React from 'react'
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
 
+import * as React from 'react'
+import { useSelector } from 'react-redux'
+import { useHistory } from 'react-router'
+
+// utils
+import { getLocale, splitStringForTag } from '../../../../../common/locale'
+
+// types
+import { PageState, WalletRoutes } from '../../../../constants/types'
+
+// style
 import {
   StyledWrapper,
   Title,
@@ -15,38 +29,62 @@ import {
   CryptoWalletsAlertTitle,
   CryptoWalletsAlertDescription
 } from './style'
+
+// components
 import { NavButton } from '../../../extension'
-import { getLocale, splitStringForTag } from '../../../../../common/locale'
 
-export interface Props {
-  onSetup: () => void
-  onRestore: () => void
-  onClickImportMetaMask: () => void
-  isMetaMaskInitialized: boolean
-  isCryptoWalletsInitialized: boolean
-}
+export const OnboardingWelcome = () => {
+  // routing
+  const history = useHistory()
 
-function OnboardingWelcome (props: Props) {
-  const { onRestore, onSetup, onClickImportMetaMask, isMetaMaskInitialized, isCryptoWalletsInitialized } = props
+  // redux
+  const {
+    isCryptoWalletsInitialized,
+    isMetaMaskInitialized
+  } = useSelector(({ page }: { page: PageState }) => page)
 
-  const onClickSettings = () => {
+  // methods
+  const onRestore = React.useCallback(() => {
+    history.push(WalletRoutes.Restore)
+  }, [])
+
+  const onClickImportMetaMask = React.useCallback(() => {
+    history.push(WalletRoutes.OnboardingImportMetaMask)
+  }, [])
+
+  const onClickSettings = React.useCallback(() => {
     chrome.tabs.create({ url: 'chrome://settings/wallet' }, () => {
       if (chrome.runtime.lastError) {
         console.error('tabs.create failed: ' + chrome.runtime.lastError.message)
       }
     })
-  }
+  }, [])
 
-  const walletAlertText = getLocale('braveWalletCryptoWalletsDescriptionTwo')
-  const { beforeTag, duringTag, afterTag } = splitStringForTag(walletAlertText)
+  const onSetup = React.useCallback(() => {
+    if (isCryptoWalletsInitialized) {
+      history.push(WalletRoutes.OnboardingImportCryptoWallets)
+      return
+    }
+    history.push(WalletRoutes.OnboardingCreatePassword)
+  }, [isCryptoWalletsInitialized])
+
+  // computed
+  const { beforeTag, duringTag, afterTag } = splitStringForTag(
+    getLocale('braveWalletCryptoWalletsDescriptionTwo')
+  )
 
   return (
     <StyledWrapper>
+
       <PageIcon />
+
       <Title>{getLocale('braveWalletWelcomeTitle')}</Title>
       <Description>{getLocale('braveWalletWelcomeDescription')}</Description>
+
       <NavButton buttonType='primary' text={getLocale('braveWalletWelcomeButton')} onSubmit={onSetup} />
+
       <RestoreButton onClick={onRestore}>{getLocale('braveWalletWelcomeRestoreButton')}</RestoreButton>
+
       {isMetaMaskInitialized &&
         <>
           <Divider />
@@ -56,6 +94,7 @@ function OnboardingWelcome (props: Props) {
           </ImportButton>
         </>
       }
+
       {isCryptoWalletsInitialized &&
         <CryptoWalletsAlertBox>
           <CryptoWalletsAlertTitle>{getLocale('braveWalletCryptoWalletsDetected')}</CryptoWalletsAlertTitle>
