+++ components/brave_wallet_ui/components/extension/connected-header/index.tsx
@@ -1,4 +1,12 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
+
+// utils
+import { getLocale } from '../../../../common/locale'
 
 // Styled Components
 import {
@@ -7,27 +15,28 @@ import {
   ActionIcon,
   ExpandIcon
 } from './style'
+
+// components
 import { WalletMorePopup } from '../../desktop'
-import { getLocale } from '../../../../common/locale'
 
 export interface Props {
   onExpand: () => void
   onClickMore: () => void
-  onClickLock: () => void
   onClickSetting: () => void
   onClickViewOnBlockExplorer: () => void
   showMore: boolean
 }
 
-const ConnectedHeader = (props: Props) => {
+export const ConnectedHeader = (props: Props) => {
   const {
-    onClickLock,
     onClickMore,
     onClickSetting,
     onExpand,
     onClickViewOnBlockExplorer,
     showMore
   } = props
+
+  // render
   return (
     <HeaderWrapper>
       <ExpandIcon onClick={onExpand} />
@@ -35,7 +44,6 @@ const ConnectedHeader = (props: Props) => {
       <ActionIcon onClick={onClickMore} />
       {showMore &&
         <WalletMorePopup
-          onClickLock={onClickLock}
           onClickSetting={onClickSetting}
           onClickViewOnBlockExplorer={onClickViewOnBlockExplorer}
         />
