+++ components/brave_wallet_ui/options/top-nav-options.ts
@@ -1,3 +1,8 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
 import { TopTabNavObjectType } from '../constants/types'
 import { getLocale } from '../../common/locale'
 
@@ -16,3 +21,5 @@ export const TopNavOptions = (): TopTabNavObjectType[] => [
   //   name: getLocale('braveWalletTopTabApps')
   // }
 ]
+
+export const TOP_NAV_OPTIONS = TopNavOptions()
