+++ components/brave_wallet_ui/page/async/wallet_page_async_handler.ts
@@ -2,6 +2,7 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at http://mozilla.org/MPL/2.0/.
+
 import getWalletPageApiProxy from '../wallet_page_api_proxy'
 import AsyncActionHandler from '../../../common/AsyncActionHandler'
 import * as WalletPageActions from '../actions/wallet_page_actions'
@@ -13,7 +14,6 @@ import {
 } from '../../constants/types'
 import {
   CreateWalletPayloadType,
-  RestoreWalletPayloadType,
   UpdateSelectedAssetType,
   ImportAccountPayloadType,
   RemoveImportedAccountPayloadType,
@@ -21,7 +21,8 @@ import {
   ViewPrivateKeyPayloadType,
   ImportAccountFromJsonPayloadType,
   ImportFromExternalWalletPayloadType,
-  ImportFilecoinAccountPayloadType
+  ImportFilecoinAccountPayloadType,
+  RestoreWalletPayloadType
 } from '../constants/action_types'
 import {
   findHardwareAccountInfo,
@@ -59,6 +60,7 @@ handler.on(WalletPageActions.restoreWallet.getType(), async (store: Store, paylo
   keyringService.notifyWalletBackupComplete()
   await refreshWalletInfo(store)
   store.dispatch(WalletPageActions.setShowIsRestoring(false))
+  store.dispatch(WalletPageActions.walletSetupComplete())
 })
 
 handler.on(WalletPageActions.showRecoveryPhrase.getType(), async (store: Store, payload: boolean) => {
