+++ components/brave_wallet_ui/stories/mock-data/user-accounts.ts
@@ -1,3 +1,8 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at http://mozilla.org/MPL/2.0/.
+
 import { UserAccountType } from '../../constants/types'
 
 export const mockUserAccounts: UserAccountType[] = [
@@ -35,3 +40,5 @@ export const mockRecoveryPhrase = [
   'glasses',
   'sound'
 ]
+
+export const mockedMnemonic = mockRecoveryPhrase.join(' ')
