+++ components/brave_wallet_ui/stories/screens/crypto-story-view.tsx
@@ -11,10 +11,7 @@ import {
   BraveWallet,
   AppsListType,
   UserAssetInfoType,
-  AccountTransactions,
   WalletAccountType,
-  UpdateAccountNamePayloadType,
-  DefaultCurrencies,
   AddAccountNavTypes
 } from '../../constants/types'
 import { TopNavOptions } from '../../options/top-nav-options'
@@ -23,34 +20,25 @@ import { SearchBar, AppList } from '../../components/shared'
 import { getLocale } from '../../../common/locale'
 import { AppsList } from '../../options/apps-list-options'
 import { filterAppList } from '../../utils/filter-app-list'
-import { PortfolioView, AccountsView } from '../../components/desktop/views'
+import { Accounts, PortfolioOverview } from '../../components/desktop/views'
 import {
   HardwareWalletConnectOpts
 } from '../../components/desktop/popup-modals/add-account-modal/hardware-wallet-connect/types'
+import WalletPageStory from '../wrappers/wallet-page-story-wrapper'
 
 export interface Props {
-  defaultCurrencies: DefaultCurrencies
   selectedNetwork: BraveWallet.NetworkInfo
   showAddModal: boolean
   userAssetList: UserAssetInfoType[]
-  transactions: AccountTransactions
-  networkList: BraveWallet.NetworkInfo[]
   accounts: WalletAccountType[]
   needsBackup: boolean
-  userVisibleTokensInfo: BraveWallet.BlockchainToken[]
-  privateKey: string
-  transactionSpotPrices: BraveWallet.AssetPrice[]
   hasImportError: boolean
   isFilecoinEnabled: boolean
   isSolanaEnabled: boolean
-  onLockWallet: () => void
   onSetImportError: (hasError: boolean) => void
   onImportAccountFromJson: (accountName: string, password: string, json: string) => void
-  onDoneViewingPrivateKey: () => void
-  onViewPrivateKey: (address: string, isDefault: boolean, coin: BraveWallet.CoinType) => void
   onRemoveAccount: (address: string, hardware: boolean, coin: BraveWallet.CoinType) => void
   onToggleAddModal: () => void
-  onUpdateAccountName: (payload: UpdateAccountNamePayloadType) => { success: boolean }
   getBalance: (address: string) => Promise<string>
   onAddHardwareAccounts: (selected: BraveWallet.HardwareWalletAccount[]) => void
   onConnectHardwareWallet: (opts: HardwareWalletConnectOpts) => Promise<BraveWallet.HardwareWalletAccount[]>
@@ -58,28 +46,17 @@ export interface Props {
   onImportFilecoinAccount: (accountName: string, privateKey: string, network: string) => void
   onCreateAccount: (name: string) => void
   onShowBackup: () => void
-  onShowVisibleAssetsModal: (value: boolean) => void
-  showVisibleAssetsModal: boolean
 }
 
 const CryptoStoryView = (props: Props) => {
   const {
-    defaultCurrencies,
     hasImportError,
-    userVisibleTokensInfo,
-    transactionSpotPrices,
-    privateKey,
     selectedNetwork,
     needsBackup,
     accounts,
-    networkList,
-    transactions,
     showAddModal,
-    showVisibleAssetsModal,
     isFilecoinEnabled,
     isSolanaEnabled,
-    onShowVisibleAssetsModal,
-    onLockWallet,
     onShowBackup,
     onCreateAccount,
     onConnectHardwareWallet,
@@ -87,17 +64,14 @@ const CryptoStoryView = (props: Props) => {
     getBalance,
     onImportAccount,
     onImportFilecoinAccount,
-    onUpdateAccountName,
     onToggleAddModal,
     onRemoveAccount,
-    onViewPrivateKey,
-    onDoneViewingPrivateKey,
     onImportAccountFromJson,
     onSetImportError
   } = props
   const [showBackupWarning, setShowBackupWarning] = React.useState<boolean>(needsBackup)
   const [showDefaultWalletBanner, setShowDefaultWalletBanner] = React.useState<boolean>(needsBackup)
-  const [selectedAccount, setSelectedAccount] = React.useState<WalletAccountType>()
+  const [, setSelectedAccount] = React.useState<WalletAccountType>()
   const [hideNav, setHideNav] = React.useState<boolean>(false)
   const [filteredAppsList, setFilteredAppsList] = React.useState<AppsListType[]>(AppsList())
   const [favoriteApps, setFavoriteApps] = React.useState<BraveWallet.AppItem[]>([
@@ -147,11 +121,6 @@ const CryptoStoryView = (props: Props) => {
     onToggleAddModal()
   }
 
-  const goBack = () => {
-    setSelectedAccount(undefined)
-    toggleNav()
-  }
-
   const onSelectAccount = (account: WalletAccountType) => {
     setSelectedAccount(account)
     toggleNav()
@@ -177,104 +146,87 @@ const CryptoStoryView = (props: Props) => {
   }
 
   return (
-    <StyledWrapper onClick={onHideMore}>
-      {!hideNav &&
-        <>
-          <TopTabNav
-            tabList={TopNavOptions()}
-            selectedTab={selectedTab}
-            onSelectTab={onSelectTab}
-            hasMoreButtons={true}
-            showMore={showMore}
-            onClickLock={onLockWallet}
-            onClickBackup={onShowBackup}
-            onClickMore={onClickMore}
-            onClickSettings={onClickSettings}
-          />
-          {showDefaultWalletBanner &&
-            <WalletBanner
-              description={getLocale('braveWalletDefaultWalletBanner')}
-              onDismiss={onDismissDefaultWalletBanner}
-              onClick={onClickSettings}
-              bannerType='warning'
-              buttonText={getLocale('braveWalletWalletPopupSettings')}
+    <WalletPageStory>
+      <StyledWrapper onClick={onHideMore}>
+        {!hideNav &&
+          <>
+            <TopTabNav
+              tabList={TopNavOptions()}
+              selectedTab={selectedTab}
+              onSelectTab={onSelectTab}
+              hasMoreButtons={true}
+              showMore={showMore}
+              onClickBackup={onShowBackup}
+              onClickMore={onClickMore}
+              onClickSettings={onClickSettings}
             />
-          }
+            {showDefaultWalletBanner &&
+              <WalletBanner
+                description={getLocale('braveWalletDefaultWalletBanner')}
+                onDismiss={onDismissDefaultWalletBanner}
+                onClick={onClickSettings}
+                bannerType='warning'
+                buttonText={getLocale('braveWalletWalletPopupSettings')}
+              />
+            }
 
-          {needsBackup && showBackupWarning &&
-            <WalletBanner
-              description={getLocale('braveWalletBackupWarningText')}
-              onDismiss={onDismissBackupWarning}
-              onClick={onShowBackup}
-              bannerType='danger'
-              buttonText={getLocale('braveWalletBackupButton')}
+            {needsBackup && showBackupWarning &&
+              <WalletBanner
+                description={getLocale('braveWalletBackupWarningText')}
+                onDismiss={onDismissBackupWarning}
+                onClick={onShowBackup}
+                bannerType='danger'
+                buttonText={getLocale('braveWalletBackupButton')}
+              />
+            }
+          </>
+        }
+        {selectedTab === 'apps' &&
+          <>
+            <SearchBar
+              placeholder={getLocale('braveWalletSearchText')}
+              action={filterList}
             />
-          }
-        </>
-      }
-      {selectedTab === 'apps' &&
-        <>
-          <SearchBar
-            placeholder={getLocale('braveWalletSearchText')}
-            action={filterList}
+            <AppList
+              list={filteredAppsList}
+              favApps={favoriteApps}
+              addToFav={addToFavorites}
+              removeFromFav={removeFromFavorites}
+              action={browseMore}
+            />
+          </>
+        }
+        {selectedTab === 'portfolio' &&
+          <PortfolioOverview />
+        }
+        {selectedTab === 'accounts' &&
+          <Accounts
+            onClickAddAccount={onClickAddAccount}
+            onRemoveAccount={onRemoveAccount}
+            onSelectAccount={onSelectAccount}
           />
-          <AppList
-            list={filteredAppsList}
-            favApps={favoriteApps}
-            addToFav={addToFavorites}
-            removeFromFav={removeFromFavorites}
-            action={browseMore}
+        }
+        {showAddModal &&
+          <AddAccountModal
+            accounts={accounts}
+            selectedNetwork={selectedNetwork}
+            onClose={onCloseAddModal}
+            onCreateAccount={onCreateAccount}
+            onImportAccount={onImportAccount}
+            onImportFilecoinAccount={onImportFilecoinAccount}
+            isFilecoinEnabled={isFilecoinEnabled}
+            isSolanaEnabled={isSolanaEnabled}
+            onConnectHardwareWallet={onConnectHardwareWallet}
+            onAddHardwareAccounts={onAddHardwareAccounts}
+            getBalance={getBalance}
+            onImportAccountFromJson={onImportAccountFromJson}
+            hasImportError={hasImportError}
+            onSetImportError={onSetImportError}
+            tab={addAccountModalTab}
           />
-        </>
-      }
-      {selectedTab === 'portfolio' &&
-        <PortfolioView
-          toggleNav={toggleNav}
-          onClickAddAccount={onClickAddAccount}
-          onShowVisibleAssetsModal={onShowVisibleAssetsModal}
-          showVisibleAssetsModal={showVisibleAssetsModal}
-        />
-      }
-      {selectedTab === 'accounts' &&
-        <AccountsView
-          defaultCurrencies={defaultCurrencies}
-          toggleNav={toggleNav}
-          accounts={accounts}
-          onClickAddAccount={onClickAddAccount}
-          onUpdateAccountName={onUpdateAccountName}
-          onRemoveAccount={onRemoveAccount}
-          onDoneViewingPrivateKey={onDoneViewingPrivateKey}
-          onViewPrivateKey={onViewPrivateKey}
-          goBack={goBack}
-          onSelectAccount={onSelectAccount}
-          privateKey={privateKey}
-          transactions={transactions}
-          transactionSpotPrices={transactionSpotPrices}
-          userVisibleTokensInfo={userVisibleTokensInfo}
-          selectedAccount={selectedAccount}
-          networkList={networkList}
-        />
-      }
-      {showAddModal &&
-        <AddAccountModal
-          accounts={accounts}
-          selectedNetwork={selectedNetwork}
-          onClose={onCloseAddModal}
-          onCreateAccount={onCreateAccount}
-          onImportAccount={onImportAccount}
-          onImportFilecoinAccount={onImportFilecoinAccount}
-          isFilecoinEnabled={isFilecoinEnabled}
-          isSolanaEnabled={isSolanaEnabled}
-          onConnectHardwareWallet={onConnectHardwareWallet}
-          onAddHardwareAccounts={onAddHardwareAccounts}
-          getBalance={getBalance}
-          onImportAccountFromJson={onImportAccountFromJson}
-          hasImportError={hasImportError}
-          onSetImportError={onSetImportError}
-          tab={addAccountModalTab}
-        />
-      }
-    </StyledWrapper>
+        }
+      </StyledWrapper>
+    </WalletPageStory>
   )
 }
