+++ components/brave_wallet_ui/stories/wallet-extension-panels.tsx
@@ -1,3 +1,8 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at http://mozilla.org/MPL/2.0/.
+
 import * as React from 'react'
 import { Provider } from 'react-redux'
 import { combineReducers, createStore } from 'redux'
@@ -450,10 +455,6 @@ export const _ConnectedPanel = (args: { locked: boolean }) => {
     }
   }
 
-  const onLockWallet = () => {
-    setWalletLocked(true)
-  }
-
   const handlePasswordChanged = (value: string) => {
     setHasPasswordError(false)
     setInputValue(value)
@@ -531,7 +532,6 @@ export const _ConnectedPanel = (args: { locked: boolean }) => {
                 selectedAccount={selectedAccount}
                 isConnected={true}
                 navAction={navigateTo}
-                onLockWallet={onLockWallet}
                 onOpenSettings={onOpenSettings}
                 originInfo={originInfo}
                 isSwapSupported={true}
