+++ components/brave_wallet_ui/utils/account-utils.ts
@@ -1,0 +1,27 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
+import { WalletAccountType } from '../constants/types'
+
+export const sortAccountsByName = (accounts: WalletAccountType[]) => {
+  return [...accounts].sort(function (a: WalletAccountType, b: WalletAccountType) {
+    if (a.name < b.name) {
+      return -1
+    }
+
+    if (a.name > b.name) {
+      return 1
+    }
+
+    return 0
+  })
+}
+
+export const groupAccountsById = (accounts: WalletAccountType[], key: string) => {
+  return accounts.reduce((result, obj) => {
+    (result[obj[key]] = result[obj[key]] || []).push(obj)
+    return result
+  }, {})
+}
